[FEAT]: update the login component structure to features/auth/presentational and features/auth/container
