+++ src/app/features/auth/container/login-container/login-container.component.html
@@ -1,0 +1,1 @@
+<app-login />
