+++ src/app/features/auth/container/login-container/login-container.component.ts
@@ -1,0 +1,11 @@
+import { ChangeDetectionStrategy, Component } from "@angular/core";
+import { LoginComponent } from "../../presentational/login/login.component";
+
+@Component({
+	selector: "app-login-container",
+	imports: [LoginComponent],
+	templateUrl: "./login-container.component.html",
+	styleUrl: "./login-container.component.scss",
+	changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class LoginContainerComponent {}
