+++ src/app/features/auth/presentational/login/login.component.html
@@ -10,11 +10,11 @@
     </div>
     <div>
       <p class="text-black">CORREO</p>
-      <app-input classInput="text-lg p-3 w-full" type="email" />
+      <app-input class="text-lg p-3 w-full" type="email" />
     </div>
     <div class="flex flex-col">
       <p class="text-black">CONTRASEÑA</p>
-      <app-input classInput="text-lg p-3 w-full" type="password" />
+      <app-input class="text-lg p-3 w-full" type="password" />
       <a class="text-black text-right font-bold" href=""
         >Olvide mi contraseña?</a
       >
